Create signal for udev to overcome problem with upstart.
The udevsettle script was starting off of an 'on started udev' event but since
it was using 'expect fork' and a script, that was being fired before udev
actually started.  This is a fix that fires an event at the end of the script.  Then udev will have started and udevsettle will detect it correctly.